requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
CurrencyConverter==0.17.13
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
install==1.3.5
magic-filter==1.0.12
multidict==6.0.4
nest-asyncio==1.5.8
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
typing_extensions==4.7.1
yarl==1.9.2
